feat(performance): extract PerformanceOrchestratorService - layer separation refactoring
- Created PerformanceOrchestratorService with singleton pattern
- Extracted business logic from route handler (217 → 66 lines, 70% reduction)
- Service handles orchestration workflow and improvement estimation
- Perfect Service Layer compliance following blueprint.md:208-209 principles
- Quality Gates: Security ✅, Lint ✅, Typecheck ✅, Build ✅
