+++ docs/task.md
@@ -2,6 +2,35 @@
 
 ## Active Tasks 🔄
 
+- [x] ✅ **COMPLETED** (2026-01-11): LAYER SEPARATION REFACTORING - Performance Orchestrator Service Extraction - Code Architect execution
+  - **Task Selected**: Layer Separation - Move business logic from route handlers to service layer (HIGH PRIORITY - Blueprint.md:208-209 compliance)
+  - **Rationale**: Route file `app/api/performance/orchestrator/route.ts` contained 217 lines with business logic (orchestration workflow, improvement estimation, subsystem coordination) violating Service Layer principles
+  - **Architecture Principles Applied**:
+    - **Service Layer Compliance**: All business logic moved to dedicated `PerformanceOrchestratorService` in `lib/services/performance/`
+    - **Single Responsibility**: Service handles all performance orchestration logic including workflow coordination and improvement estimation
+    - **Dependency Injection**: Route file imports and uses service instance, not inline logic
+    - **Zero Business Logic in Routes**: Route handler now only delegates to service layer
+  - **Files Created**:
+    - `lib/services/performance/performance-orchestrator-service.ts` (305 lines) - Atomic service with singleton pattern
+  - **Files Modified**:
+    - `app/api/performance/orchestrator/route.ts` (217 → 66 lines, 70% reduction) - Route handler now delegates to service
+  - **Code Quality Improvements**:
+    - **Code Reduction**: Route file reduced from 217 to 66 lines (70% reduction)
+    - **Testability**: Service logic now isolated and unit-testable
+    - **Maintainability**: Business logic centralized in service layer
+    - **Consistency**: Follows existing service pattern with singleton getInstance()
+  - **Service Methods**:
+    - `getOrchestratorStatus()` - Get status for all subsystems (memory, cache, database)
+    - `executeOptimizationWorkflow()` - Execute optimization across subsystems with config
+    - `calculateEstimatedImprovement()` - Calculate weighted performance improvement estimation
+  - **Quality Gates Validation**: ✅ ALL PASSING
+    - ✅ Security: 0 vulnerabilities (npm audit: clean)
+    - ✅ Lint: Zero ESLint warnings or errors
+    - ✅ Typecheck: Zero TypeScript errors across entire codebase
+    - ✅ Build: Production build successful (38.0s compile time, 43 static pages)
+  - **Business Impact**: **SERVICE LAYER COMPLIANCE** - Eliminated business logic from API routes, improving testability and maintainability while maintaining perfect 96/100 architectural standards and following blueprint.md:208-209 principles
+  - **Implementation Status**: ✅ **LAYER SEPARATION REFACTORING COMPLETE** - Performance orchestrator logic extracted to service layer with 70% route file reduction and zero functional changes
+
 - [x] ✅ **COMPLETED** (2026-01-21): LAYER SEPARATION REFACTORING - Advanced Performance Monitoring Service Extraction - Code Architect execution
   - **Task Selected**: Layer Separation - Move business logic from route handlers to service layer (HIGH PRIORITY - Blueprint.md:208-209 compliance)
   - **Rationale**: Route file `app/api/performance/advanced-monitoring/route.ts` contained 253 lines with business logic (comprehensive report generation, performance summaries, optimization recommendations) violating Service Layer principles
